Handle p4dctl silliness
Unlike normal service management tools, p4dctl returns an error code (1)
if the service is not running, even it is appropriately configured. The
same result is returned when the service is misconfigured for any reason
(missing mount dir, missing user, incorrect path, etc.). Handle it by
explicitly checking the output of the command for the expected string if
the service is set up but simply not started yet.